Add binary-search methods
Add methods to collections to perform binary searches.  All the methods assume the collection is sorted along the given predicate, or simply in non-decreasing order for the overloads that default the predicate to the standard less-than operator.  The variants return: the quickest match, the earliest match, one past the latest match, and the index range for all matches.