Add method for the first divergence between sequences
Add a method to sequences that takes another sequence and an equivalence predicate to return the corresponding elements of where the sequences initially don't match.  If at least one sequence runs out before a mismatch, its corresponding return value is NIL.  Add an overload that defaults comparison to the standard equality operator.