Add method for the first divergence between collections
Add a method to collections that takes another collection and an equivalence predicate to return the indexes for the corresponding elements of where the collections initially don't match. If at least one collection runs out before a mismatch, its corresponding return value is its source's past-the-end index. Add an overload that defaults comparison to the standard equality operator.