Add method to check sequence overlap
Add method to sequences that takes another sequence and a comparison predicate, assumes both sequences are sorted along that predicate, and reports how much the sequences overlap in element values.  Add an overload that defaults the comparison to the standard less-than operator.  Add a type to express the form of set overlap, making optional distinction to degenerate sources.