Add methods to merge sorted sequences
Add a type describing the subsets that can be taken of the merger of two sets.

Add a type that is a sequence that takes two sequences, a predicate that both sources are sorted along, and a merger-subset state, and vends the seleected subset of the merger which is also sorted.  Add an extension method to sequences that takes another sequence, ordering predicate, and subset selection and returns an array with the merged sequence.  Add lazy variants of this method.  Add overloads that default the ordering predicate to the standard less-than operator.