Add method listing element deltas lazily
Add a type for a sequence that wraps another and caches an evaluation closure, vending the results of calling that closure for each pair of adjacent elements in order.

Add a method to lazy sequences that overloads the method to return the differences between adjacent elements, returning instances of the new sequence type.