Add specializations for listing element deltas
Add variants of the method for sequences that takes an evaluation closure and returns a lazy sequence holding the results of calling that closure for each pair of adjacent elements in order.  These variants lock the closure to an operator or method from a given protocol.  The variants use the subtraction, wrapping subtraction, and striding distance operations.