Add methods for in-place scan and un-scan
Add a method to element-mutable collections that takes a closure that can fuse two element values and applies it to the collection, i.e. progressive reduce, a.k.a. scan.  Add another method that takes a closure that can separate an element value from another and applies it to a collection, doing the inverse of the first method.