Add method to copy elements
Add method to element-mutable collections that takes a sequence of the same element type to read its elements and assign them on top of the collection's current elements.  The copying stops upon reaching the end of the shorter sequence.  The method returns two items.  The first is one index past the last destination element written over.  The second is an iterator for the elements of the source sequence that were not used for copying.