Add method to copy elements backwards
Add method to element-mutable bidirectional collections that takes a bidirectional collection with the same element type to read its last elements and assign them on top of the receiver's last elements.  The copying is done backwards and stops upon at least one of the collections reaching their beginning (whichever hits it first).  The method returns two items: the starting indices of each collection's suffix that participated in the copying.