Add methods to internally copy elements
Add method to element-mutable collections that takes range expressions for two subsequences of the receiver and copies elements from one subsequence on top of the elements for the other.  The copying is done forwards, targeting the subsequences' prefixes until at least one subsequence runs out of elements.  Add an overload for bidirectional collections that does its copies in reverse over the subsequences' suffixes instead.  The methods return two items; range expressions bounding the parts of the subsequences that actually participated in copying.