Improve documentation; rename methods
Improve the summary descriptions for the methods of element-mutable collections that copy over elements, but only for the overloads where the source is a sequence or collection.  They clarify what parts of the receiver and source are touched.

Rename the "copyOntoSuffix" methods to use "copy" by itself as the base names, and use new parameter labels to differentiate the overloads.  Now all of the element-copying methods use the same base name.