Simplify copying from a sequence
Move the secret implementation methods to the end of their file. Change the methods that copy from a sequence to return only the index of the first untouched element of the receiver. (Use the iterator-based overload to track unread elements from the source.)

Move the location of the revised tests in their file. Update the documentation, including the prime example.