Add function to test the inclusion of one sorted sequence into another
Add a function for a sequence that takes another sequence and a predicate, assumes that both sequences are sorted to that predicate, and checks if the argument sequence is a subset of the receiver.

Add a function that takes the above function, but defaults the predicate to the less-than operator.

Add tests and documentation.