Remove the error type in a public declaration
The sequence to merge other sequences needs to specify the ordering predicate's error type. But public use of that name never needs that error type. Rename the merging sequence to a un-public name. Add an alias from that un-public name to the old public name but removing the error type.