Add labels to the tuples produced by `chunked(on:)`
Label the tuple produced by `chunked(on:)` with `subject` and
`chunk`. This will improve the legibility of code downstream of
the said function. For example, when providing an identifier
for SwiftUI's `ForEach`:

```
-ForEach(chunkedByDate, id: \.0) { date, items in
+ForEach(chunkedByDate, id: \.subject) { date, items in
```

This aligns `ChunkedOnCollection` with the `IndexedCollection`,
which specifies `index` and `element` as well as the built-in
`EnumeratedSequence` (specifying `offset` and `element`).

The introduction of the label will have no effect on existing code,
because the index-based address of the tuple component will remain
unchanged and absent labels will be inferred.